+++ backend/app/routers/auth/__init__.py
@@ -1,0 +1,9 @@
+"""
+auth ルーターパッケージ。
+
+外部から `router` をインポートできるようにエクスポートする。
+"""
+
+from .endpoints import router
+
+__all__ = ["router"]
