+++ backend/app/services/intelligence/github/__init__.py
@@ -1,0 +1,6 @@
+"""
+GitHub 分析サブパッケージ。
+
+api_client: GitHub REST API 呼び出し
+repo_analyzer: 言語 ratio 算出・フレームワーク検出・依存関係パース
+"""
